added session support

+++ DeactivationService/Controllers/DeactivationController.cs
@@ -7,6 +7,7 @@ using Microsoft.AspNetCore.Mvc;
 using DeactivationService.Services;
 using Microsoft.Extensions.Configuration;
 using DeactivationService.Models;
+using Microsoft.Extensions.Caching.Memory;
 
 namespace DeactivationService.Controllers
 {
@@ -15,11 +16,13 @@ namespace DeactivationService.Controllers
     [ApiController]
     public class DeactivationController : ControllerBase
     {
-		DeviceDeactivationService deactivationService;
+		DeviceDeactivationService	deactivationService;
+		SessionService				sessionService;
 
-		public DeactivationController(IConfiguration configuration)
+		public DeactivationController(IConfiguration configuration, IMemoryCache memoryCache)
 		{
 			deactivationService = new DeviceDeactivationService(configuration);
+			sessionService		= new SessionService(configuration, memoryCache);
 		}
 
 		[HttpGet]
@@ -27,6 +30,12 @@ namespace DeactivationService.Controllers
 		public IActionResult GetReport(int companyId, int page, int pageSize, 
 			bool onlyPending, string sortColumn, bool sortAsc, string filterColumn, string filterValue)
 		{
+			string authToken = Request.Headers["x-access-token"];
+			SessionService.SESSION_STATE sessionState = sessionService.CheckSession(authToken, companyId);
+
+			if(sessionState == SessionService.SESSION_STATE.INVALID)
+				return StatusCode(401);
+
 			string[] validSortColumns = { "requestDate", "status", "reason", "completedDate", "username" };
 			if(!validSortColumns.Contains(sortColumn))
 			{
@@ -38,8 +47,14 @@ namespace DeactivationService.Controllers
 
 		[HttpGet]
 		[ActionName("getRequest")]
-		public IActionResult GetRequest(string requestId)
+		public IActionResult GetRequest(string requestId, int companyId)
 		{
+			string authToken = Request.Headers["x-access-token"];
+			SessionService.SESSION_STATE sessionState = sessionService.CheckSession(authToken, companyId);
+
+			if(sessionState == SessionService.SESSION_STATE.INVALID)
+				return StatusCode(401);
+
 			DeactivateRequest report = deactivationService.GetDeactivationRequest(requestId);
 			return new ObjectResult(report);
 		}
@@ -48,6 +63,12 @@ namespace DeactivationService.Controllers
 		[ActionName("getStatusSummary")]
 		public IActionResult GetStatusSummary(int companyId)
 		{
+			string authToken = Request.Headers["x-access-token"];
+			SessionService.SESSION_STATE sessionState = sessionService.CheckSession(authToken, companyId);
+
+			if(sessionState == SessionService.SESSION_STATE.INVALID)
+				return StatusCode(401);
+
 			List<StatusSummary> summary = deactivationService.GetStatusSummary(companyId);
 
 			return new ObjectResult(summary);
@@ -59,7 +80,11 @@ namespace DeactivationService.Controllers
 		{
 			// Verify the user request
 			string authToken = Request.Headers["x-access-token"];
-			request.userId = 1298352; // Replace this with converted session Id
+			SessionService.SESSION_STATE state = sessionService.CheckSession(authToken, request.cid);
+			request.userId = sessionService.GetUserId(authToken);
+
+			if(state == SessionService.SESSION_STATE.INVALID || request.userId == -1)
+				return StatusCode(401);
 
 			string requestId = deactivationService.Deactivate(request);
 
@@ -72,8 +97,15 @@ namespace DeactivationService.Controllers
 		{
 			if (device == null || device.cid == null || device.dsn == null || device.requestId == null)
 			{
-				return new ObjectResult(false);
+				return StatusCode(400);
 			}
+
+			string authToken = Request.Headers["x-access-token"];
+			SessionService.SESSION_STATE state = sessionService.CheckSession(authToken, Convert.ToInt32(device.cid));
+
+			// TODO: Validate status vs session
+			if(state == SessionService.SESSION_STATE.INVALID)
+				return StatusCode(401);
 
 			bool response = deactivationService.RemoveDeviceFromRequest(Convert.ToInt32(device.cid), Convert.ToInt32(device.dsn), device.requestId);
 			return new ObjectResult(response);
@@ -83,9 +115,16 @@ namespace DeactivationService.Controllers
 		[ActionName("cancelRequest")]
 		public IActionResult CancelRequest([FromBody] DeactivateRequest request)
 		{
+			string authToken = Request.Headers["x-access-token"];
+			SessionService.SESSION_STATE state = sessionService.CheckSession(authToken, Convert.ToInt32(request.cid));
+
+			// TODO: Validate status vs session
+			if(state == SessionService.SESSION_STATE.INVALID)
+				return StatusCode(401);
+
 			if (request.requestId == null || request.requestId == null)
 			{
-				return new ObjectResult(false);
+				return StatusCode(400);
 			}
 
 			bool response = deactivationService.CancelRequest(request.requestId.ToString());
@@ -98,9 +137,14 @@ namespace DeactivationService.Controllers
 		public IActionResult UpdateRequest([FromBody] DeactivateRequest request)
 		{
 			if (request == null)
-			{
-				return new ObjectResult(false);
-			}
+				return StatusCode(400);
+
+			string authToken = Request.Headers["x-access-token"];
+			SessionService.SESSION_STATE state = sessionService.CheckSession(authToken, Convert.ToInt32(request.cid));
+
+			// TODO: Validate status vs session
+			if(state == SessionService.SESSION_STATE.INVALID)
+				return StatusCode(401);
 
 			bool response = deactivationService.UpdateRequest(
 				request.requestId.ToString(), 
@@ -118,6 +162,13 @@ namespace DeactivationService.Controllers
 		[ActionName("updateRequestStatus")]
 		public StatusCodeResult UpdateRequestStatus(UpdateStatus update)
 		{
+			string authToken = Request.Headers["x-access-token"];
+			SessionService.SESSION_STATE state = sessionService.CheckSession(authToken, Convert.ToInt32(update.cid));
+
+			// TODO: Validate status vs session
+			if(state == SessionService.SESSION_STATE.INVALID)
+				return StatusCode(401);
+
 			bool success = this.deactivationService.UpdateRequestStatus(update);
 
 			return (success)? StatusCode(200) : StatusCode(400);
@@ -127,6 +178,13 @@ namespace DeactivationService.Controllers
 		[ActionName("updateRequestReason")]
 		public StatusCodeResult UpdateRequestReason(UpdateReason update)
 		{
+			string authToken = Request.Headers["x-access-token"];
+			SessionService.SESSION_STATE state = sessionService.CheckSession(authToken, Convert.ToInt32(update.cid));
+
+			// TODO: Validate status vs session
+			if(state == SessionService.SESSION_STATE.INVALID)
+				return StatusCode(401);
+
 			bool success = this.deactivationService.UpdateRequestReason(update);
 
 			return (success)? StatusCode(200) : StatusCode(400);
@@ -136,6 +194,15 @@ namespace DeactivationService.Controllers
 		[ActionName("updateDeviceFees")]
 		public IActionResult UpdateDeviceFees([FromBody] List<Device> deviceList)
 		{
+			if(deviceList.Count == 0)
+				return StatusCode(400);
+
+			string authToken = Request.Headers["x-access-token"];
+			SessionService.SESSION_STATE state = sessionService.CheckSession(authToken, Convert.ToInt32(deviceList[0].cid));
+
+			if(state != SessionService.SESSION_STATE.MANAGER)
+				return StatusCode(401);
+
 			List<bool> response = deactivationService.UpdateDeviceFees(deviceList);
 			return new ObjectResult(response);
 		}
